requirements.txt
langchain
streamlit
google-cloud-bigquery